Use universal_html and add cursor position
