+++ CHANGELOG.md
@@ -1,19 +1,25 @@
-## [1.0.2] - 3/14/21
+## 1.1.0
+
+* Added support for cursor position in on dropped callback.
+* Replaced `dart:html` imports with `universal_html` to allow for compilation
+  on other platforms.
+
+## 1.0.2
 
 * Enable multiple drop zones
 
-## [1.0.1] - 2/22/21
+## 1.0.1
 
 * Null Safety fixes
 
-## [1.0.0] - 2/22/21
+## 1.0.0
 
 * Null Safety
 
-## [0.0.2] - 2/22/21
+## 0.0.2
 
 * Add readme, example, and demo gif
 
-## [0.0.1] - 2/22/21
+## 0.0.1
 
 * initial release.
