+++ lib/src/dispatcher.dart
@@ -1,6 +1,6 @@
 import 'dart:async';
-// ignore: avoid_web_libraries_in_flutter
-import 'dart:html' as html;
+
+import 'package:universal_html/html.dart' as html;
 
 class Dispatcher {
   static final shared = Dispatcher._internal();
